backend/pyproject.toml
[project]
name = "almabase-assignment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.31.0",
    "chromadb>=0.5.0",
    "email-validator>=2.1.0",
    "fastapi>=0.135.1",
    "bcrypt>=4.0.0,<4.1.0",
    "httpx>=0.27.0",
    "passlib[bcrypt]>=1.7.4",
    "psycopg[binary]>=3.3.3",
    "pypdf>=4.0.0",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.4.0",
    "python-multipart>=0.0.9",
    "sqlalchemy>=2.0.48",
    "uvicorn>=0.41.0",
]
